requirements.txt
asttokens==3.0.0
bidict==0.23.1
contourpy==1.3.0
cycler==0.12.1
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.0
fonttools==4.59.0
h11==0.16.0
importlib_resources==6.5.2
ipython==8.18.1
jedi==0.19.2
kiwisolver==1.4.7
matplotlib==3.9.4
matplotlib-inline==0.1.7
numpy==2.0.2
opencv-contrib-python==4.12.0.88
packaging==25.0
pandas==2.3.1
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
prompt_toolkit==3.0.51
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2025.2
simple-websocket==1.1.0
six==1.17.0
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
wcwidth==0.2.13
wsproto==1.2.0
zipp==3.23.0
